refactor: move  server time offset computation to callmanager and defferred
