+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/UserSessionScope.kt
@@ -995,6 +995,7 @@ class UserSessionScope internal constructor(
     private val callRepository: CallRepository by lazy {
         CallDataSource(
             callApi = authenticatedNetworkContainer.callApi,
+            serverTimeApi = authenticatedNetworkContainer.serverTimeApi,
             qualifiedIdMapper = qualifiedIdMapper,
             callDAO = userStorage.database.callDAO,
             conversationRepository = conversationRepository,
@@ -1319,7 +1320,6 @@ class UserSessionScope internal constructor(
 
     private val eventRepository: EventRepository = EventDataSource(
         notificationApi = authenticatedNetworkContainer.notificationApi,
-        serverTimeApi = authenticatedNetworkContainer.serverTimeApi,
         metadataDAO = userStorage.database.metadataDAO,
         eventDAO = userStorage.database.eventDAO,
         currentClientId = clientIdProvider,
