+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/sync/incremental/EventGathererTest.kt
@@ -275,12 +275,6 @@ class EventGathererTest {
             }.returns(either)
         }
 
-        suspend fun withFetchServerTimeReturning(time: String?) = apply {
-            coEvery {
-                eventRepository.fetchServerTime()
-            }.returns(time)
-        }
-
         suspend fun withLastEventIdReturning(either: Either<StorageFailure, String>) = apply {
             coEvery {
                 eventRepository.lastSavedEventId()
